requirements.txt
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
requests==2.32.3
pytz==2024.2